refactor(files): Separate backend from frontend
