+++ backend/src/main/java/com/codecool/solarwatch/client/SolarEventClient.java
