+++ backend/src/main/java/com/codecool/solarwatch/client/response/GeoResponse.java
