+++ backend/src/main/java/com/codecool/solarwatch/client/response/SolarEventResponse.java
