+++ backend/src/main/java/com/codecool/solarwatch/configuration/ClientConfig.java
