+++ backend/src/main/java/com/codecool/solarwatch/controller/SolarEventController.java
