+++ backend/src/main/java/com/codecool/solarwatch/exception/ApiError.java
