+++ backend/src/main/java/com/codecool/solarwatch/exception/CityExistsException.java
