+++ backend/src/main/java/com/codecool/solarwatch/exception/GlobalExceptionHandler.java
@@ -1,0 +1,93 @@
+package com.codecool.solarwatch.exception;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MissingServletRequestParameterException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
+
+import java.time.LocalDateTime;
+
+@RestControllerAdvice
+@Slf4j
+public class GlobalExceptionHandler {
+
+  @ExceptionHandler(ResourceNotFoundException.class)
+  public ResponseEntity<ApiError> handleResourceNotFound(ResourceNotFoundException e) {
+    log.warn(e.getMessage());
+
+    ApiError error = ApiError.builder()
+            .timestamp(LocalDateTime.now())
+            .status(HttpStatus.NOT_FOUND.value())
+            .error(HttpStatus.NOT_FOUND.toString())
+            .message(e.getMessage())
+            .build();
+
+    return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
+  }
+
+  @ExceptionHandler(MethodArgumentTypeMismatchException.class)
+  public ResponseEntity<ApiError> handleMethodArgumentTypeMismatch(MethodArgumentTypeMismatchException e) {
+    log.warn("Parameter mismatch. Param: '{}', Value: '{}', Message: {}",
+            e.getName(), e.getValue(), e.getMessage());
+
+    ApiError error = ApiError.builder()
+            .timestamp(LocalDateTime.now())
+            .status(HttpStatus.BAD_REQUEST.value())
+            .error(HttpStatus.BAD_REQUEST.toString())
+            .build();
+
+    if ("date".equals(e.getName())) {
+      error.setMessage("Invalid Date. Please use YYYY-MM-DD format.");
+      return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+    }
+
+    error.setMessage(e.getMessage());
+    return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+  }
+
+  @ExceptionHandler(MissingServletRequestParameterException.class)
+  public ResponseEntity<ApiError> handleMissingRequestParam(MissingServletRequestParameterException e) {
+    log.warn("Missing required parameter: '{}'", e.getParameterName());
+
+    ApiError error = ApiError.builder()
+            .timestamp(LocalDateTime.now())
+            .status(HttpStatus.BAD_REQUEST.value())
+            .error(HttpStatus.BAD_REQUEST.toString())
+            .message("Missing parameter: " + e.getParameterName())
+            .build();
+
+    return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+  }
+
+  @ExceptionHandler(UserExistsException.class)
+  public ResponseEntity<ApiError> handleUserExists(UserExistsException e) {
+    log.warn(e.getMessage());
+
+    ApiError error = ApiError.builder()
+            .timestamp(LocalDateTime.now())
+            .status(HttpStatus.CONFLICT.value())
+            .error(HttpStatus.CONFLICT.toString())
+            .message(e.getMessage())
+            .build();
+
+    return new ResponseEntity<>(error, HttpStatus.CONFLICT);
+  }
+
+  @ExceptionHandler(Exception.class)
+  public ResponseEntity<ApiError> handleGlobalException(Exception e) {
+    log.error("An unexpected internal server error occurred.", e);
+
+    ApiError error = ApiError.builder()
+            .timestamp(LocalDateTime.now())
+            .status(HttpStatus.INTERNAL_SERVER_ERROR.value())
+            .error(HttpStatus.INTERNAL_SERVER_ERROR.toString())
+            .message("Something went wrong. Please try again.")
+            .build();
+
+    return new ResponseEntity<>(error, HttpStatus.INTERNAL_SERVER_ERROR);
+  }
+
+}
