+++ backend/src/main/java/com/codecool/solarwatch/exception/ResourceNotFoundException.java
