+++ backend/src/main/java/com/codecool/solarwatch/exception/SolarEventExistsException.java
