+++ backend/src/main/java/com/codecool/solarwatch/exception/UserExistsException.java
