+++ backend/src/main/java/com/codecool/solarwatch/mapper/SolarEventMapper.java
