+++ backend/src/main/java/com/codecool/solarwatch/model/SolarEvent.java
