+++ backend/src/main/java/com/codecool/solarwatch/model/dto/UserAuthDTO.java
