+++ backend/src/main/java/com/codecool/solarwatch/model/dto/UserResponseDTO.java
