+++ backend/src/main/java/com/codecool/solarwatch/model/user/Role.java
