+++ backend/src/main/java/com/codecool/solarwatch/model/user/SolarWatchUser.java
