+++ backend/src/main/java/com/codecool/solarwatch/security/WebSecurityConfig.java
