+++ backend/src/main/java/com/codecool/solarwatch/service/SolarEventService.java
