+++ backend/src/main/java/com/codecool/solarwatch/service/UserService.java
