+++ backend/src/test/java/com/codecool/solarwatch/controller/SolarEventControllerTest.java
