+++ backend/src/test/java/com/codecool/solarwatch/service/SolarEventServiceTest.java
